TrailSong/requirements.txt
geopy==2.4.1
lxml==5.3.0
matplotlib==3.9.2
numpy==2.0.1
scipy==1.14.0
SRTM.py==0.3.7
